TXM-5692 - Generate magic Number as a long

+++ Gaming/Aristocrat.Monaco.Gaming/Progressives/IMysteryProgressiveProvider.cs
@@ -12,7 +12,7 @@
         /// </summary>
         /// <param name="progressiveLevel">The progressive level for which to generate the magic number</param>
         /// <returns>The magic number generated</returns>
-        decimal GenerateMagicNumber(IViewableProgressiveLevel progressiveLevel);
+        long GenerateMagicNumber(IViewableProgressiveLevel progressiveLevel);
 
         /// <summary>
         ///     Retrieves the magic number for a progressive level
@@ -20,7 +20,7 @@
         /// <param name="progressiveLevel">The progressive level for which to retrieve the magic number</param>
         /// <param name="magicNumber">The magic number for the level, if found</param>
         /// <returns>True if the progressive level has a magic number associated, otherwise false</returns>
-        bool TryGetMagicNumber(IViewableProgressiveLevel progressiveLevel, out decimal magicNumber);
+        bool TryGetMagicNumber(IViewableProgressiveLevel progressiveLevel, out long magicNumber);
 
         /// <summary>
         ///     Determines if the progressive level has met the requirements for triggering mystery jackpot
